+++ app/src/main/java/vuong/hx/tayduky/ui/fragments/dialogs/ChallengeToolsDialogFragment.java
@@ -75,7 +75,7 @@ public class ChallengeToolsDialogFragment extends DialogFragment
 
         mChallenge = (Challenge) getArguments().getSerializable("challenge");
 
-        mPresenter.loadChallengeTools(mChallenge.getId());
+        mPresenter.loadChallengeTools((int) mChallenge.getId());
         return view;
     }
